requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
aiofiles
python-dotenv
openai
aliyun-python-sdk-core
psycopg2-binary
sqlalchemy[asyncio]
asyncpg
alembic
greenlet
wave
numpy
pydub
ffmpeg-python
geopy